improve year and rating validations

+++ src/schemas/movie.schema.ts
@@ -3,18 +3,18 @@ import { z } from 'zod';
 export const moviePostSchema = z.object({
   title: z.string().min(1, { message: 'Title is required' }).max(100),
   country: z.string().min(1, { message: 'Invalid country' }).max(60).optional(),
-  year: z.number().min(1900, { message: 'Invalid year' }).optional(),
+  year: z.number().int().min(1900, { message: 'Invalid year' }).optional(),
 });
 
 export const moviePutSchema = z.object({
   title: z.string().min(1).max(100).optional(),
   country: z.string().min(1, { message: 'Invalid country' }).max(60).optional(),
-  year: z.number().min(1900, { message: 'Invalid year' }).optional(),
+  year: z.number().int().min(1900, { message: 'Invalid year' }).optional(),
 });
 
 export const movieReviewPostSchema = z.object({
   movieId: z.string().uuid(),
-  rating: z.number().min(1).max(5),
+  rating: z.number().int().min(1).max(5),
   review: z
     .string()
     .min(1, { message: 'Please provide a review' })
@@ -23,7 +23,7 @@ export const movieReviewPostSchema = z.object({
 });
 
 export const movieReviewPutSchema = z.object({
-  rating: z.number().min(1).max(5).optional(),
+  rating: z.number().int().min(1).max(5).optional(),
   review: z
     .string()
     .min(1, { message: 'Please provide a review' })
